moving interpreter package to plugin (#32997)
