--- packages/kbn-interpreter/public/package.json
@@ -1,5 +1,0 @@
-{
-  "private": true,
-  "main": "../target/public/index.js",
-  "jsnext:main": "../src/public/index.js"
-}
