--- packages/kbn-interpreter/server/package.json
@@ -1,5 +1,0 @@
-{
-  "private": true,
-  "main": "../target/server/index.js",
-  "jsnext:main": "../src/server/index.js"
-}
