+++ packages/kbn-interpreter/src/common/index.js
@@ -17,15 +17,10 @@
  * under the License.
  */
 
-export { FunctionsRegistry } from './lib/functions_registry';
-export { TypesRegistry } from './lib/types_registry';
-export { createError } from './interpreter/create_error';
-export { interpreterProvider } from './interpreter/interpret';
-export { serializeProvider } from './lib/serialize';
 export { fromExpression, toExpression, safeElementFromExpression } from './lib/ast';
 export { Fn } from './lib/fn';
 export { getType } from './lib/get_type';
-export { castProvider } from './interpreter/cast';
+export { castProvider } from './lib/cast';
 export { parse } from './lib/grammar';
 export { getByAlias } from './lib/get_by_alias';
 export { Registry } from './lib/registry';
