+++ packages/kbn-interpreter/src/common/lib/cast.js
