--- packages/kbn-interpreter/src/public/index.js
@@ -1,28 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-export { initializeInterpreter } from './interpreter';
-export { RenderFunctionsRegistry } from './render_functions_registry';
-export { registries } from './registries';
-import { registries } from './registries';
-import { registryFactory } from '../common';
-
-// Expose kbnInterpreter.register(specs) and kbnInterpreter.registries() globally so that plugins
-// can register without a transpile step.
-global.kbnInterpreter = Object.assign(global.kbnInterpreter || {}, registryFactory(registries));
