+++ src/legacy/core_plugins/interpreter/common/interpreter/create_error.js
