+++ src/legacy/core_plugins/interpreter/common/interpreter/interpret.js
@@ -19,10 +19,7 @@
 
 import clone from 'lodash.clone';
 import { each, keys, last, mapValues, reduce, zipObject } from 'lodash';
-import { getType } from '../lib/get_type';
-import { fromExpression } from '../lib/ast';
-import { getByAlias } from '../lib/get_by_alias';
-import { castProvider } from './cast';
+import { getType, fromExpression, getByAlias, castProvider } from '@kbn/interpreter/common';
 import { createError } from './create_error';
 
 export function interpreterProvider(config) {
