+++ src/legacy/core_plugins/interpreter/common/types/boolean.js
