+++ src/legacy/core_plugins/interpreter/common/types/error.js
