+++ src/legacy/core_plugins/interpreter/common/types/filter.js
