+++ src/legacy/core_plugins/interpreter/common/types/image.js
