+++ src/legacy/core_plugins/interpreter/common/types/kibana_context.js
