+++ src/legacy/core_plugins/interpreter/common/types/null.js
