+++ src/legacy/core_plugins/interpreter/common/types/number.js
