+++ src/legacy/core_plugins/interpreter/common/types/pointseries.js
