+++ src/legacy/core_plugins/interpreter/common/types/register.js
