+++ src/legacy/core_plugins/interpreter/common/types/render.js
