+++ src/legacy/core_plugins/interpreter/common/types/shape.js
