+++ src/legacy/core_plugins/interpreter/common/types/string.js
