+++ src/legacy/core_plugins/interpreter/common/types/style.js
