+++ src/legacy/core_plugins/interpreter/public/functions/clog.js
