+++ src/legacy/core_plugins/interpreter/public/functions/index.js
@@ -17,6 +17,7 @@
  * under the License.
  */
 
+import { clog } from './clog';
 import { esaggs } from './esaggs';
 import { kibana } from './kibana';
 import { kibanaContext } from './kibana_context';
@@ -35,6 +36,6 @@ import { vislib } from './vislib';
 import { visualization } from './visualization';
 
 export const functions = [
-  esaggs, kibana, kibanaContext, vega, timelionVis, tsvb, kibanaMarkdown, inputControlVis,
+  clog, esaggs, kibana, kibanaContext, vega, timelionVis, tsvb, kibanaMarkdown, inputControlVis,
   metric, kibanaPie, regionmap, tilemap, kibanaTable, tagcloud, vislib, visualization
 ];
