+++ src/legacy/core_plugins/interpreter/public/lib/batched_fetch.js
