+++ src/legacy/core_plugins/interpreter/public/lib/consts.js
