+++ src/legacy/core_plugins/interpreter/public/lib/create_handlers.js
