+++ src/legacy/core_plugins/interpreter/public/lib/interpreter.test.js
@@ -20,11 +20,11 @@
 import { FUNCTIONS_URL } from './consts';
 import { initializeInterpreter } from './interpreter';
 
-jest.mock('../common/interpreter/interpret', () => ({
+jest.mock('../../common/interpreter/interpret', () => ({
   interpreterProvider: () => () => ({}),
 }));
 
-jest.mock('../common/lib/serialize', () => ({
+jest.mock('../../common/serialize', () => ({
   serializeProvider: () => ({ serialize: () => ({}) }),
 }));
