+++ src/legacy/core_plugins/interpreter/public/lib/render_function.js
