+++ src/legacy/core_plugins/interpreter/public/registries/functions_registry.js
@@ -17,21 +17,6 @@
  * under the License.
  */
 
-import { get, identity } from 'lodash';
-import { getType } from './get_type';
+import { FunctionsRegistry } from '../../common/functions_registry';
 
-export function serializeProvider(types) {
-  return {
-    serialize: provider('serialize'),
-    deserialize: provider('deserialize'),
-  };
-
-  function provider(key) {
-    return context => {
-      const type = getType(context);
-      const typeDef = types[type];
-      const fn = get(typeDef, key) || identity;
-      return fn(context);
-    };
-  }
-}
+export const functionsRegistry = new FunctionsRegistry();
