+++ src/legacy/core_plugins/interpreter/server/routes/server_functions.js
@@ -18,7 +18,7 @@
  */
 
 import Boom from 'boom';
-import { serializeProvider } from '@kbn/interpreter/common';
+import { serializeProvider } from '../../common/serialize';
 import { API_ROUTE } from '../../common/constants';
 import { createHandlers } from '../lib/create_handlers';
 import Joi from 'joi';
